requirements.txt
# Python 3.10+ 推奨。別PCでは venv を作り、このファイルと requirements-dev.txt をインストールすること。
# システム依存: Node 18+ / npm（サイトビルド）、Claude Code CLI（npm install -g @anthropic-ai/claude-code + claude auth login）
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
openai==1.12.0
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
gitpython==3.1.40
pygithub==2.1.1
# Vercel API は modules/vercel_client.py で requests を使用（PyPI の vercel パッケージは未使用）
beautifulsoup4==4.12.3
# Python 3.12+ ではバイナリホイールを優先（ソースビルド失敗・遅延を避ける）
lxml>=5.3.0,<6
jinja2==3.1.3
aiohttp==3.9.3
# asyncio は標準ライブラリのため PyPI パッケージは不要（古い asyncio パッケージは Python 3.10+ で競合・ビルド失敗の原因になる）
